--- my-react-app/src/app/screens/Login/components/Fields/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export const customInput = props => {
-  return (
-    <div>
-      <label>{props.label}</label>
-      <input type={props.type} />
-    </div>
-  );
-};
