+++ my-react-app/src/app/screens/Login/utils.js
@@ -1,0 +1,3 @@
+export const minLength = value => (value && value.length < 8 ? ` Must be 8 characters or more` : undefined);
+
+export const required = value => (value ? undefined : 'Required');
