+++ my-react-app/src/index.js
@@ -2,14 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import { store } from './redux/store';
+import store from './redux/store.js';
 import './scss/index.scss';
 import registerServiceWorker from './registerServiceWorker';
-import App from './App';
+import LoginFormContainer from './app/screens/Login/index.js';
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <LoginFormContainer />
   </Provider>,
   document.getElementById('root')
 );
